test: Write tests for dict template
