--- django_slack_bot/utils/validators.py
@@ -1,19 +1,0 @@
-"""Validators for handy."""
-from __future__ import annotations
-
-from typing import Any
-
-from django.core.exceptions import ValidationError
-from django.utils.translation import gettext_lazy as _
-
-
-def dict_template_validator(value: Any) -> None:
-    """Validate given value is valid dictionary template."""
-    if value is None:  # No-op template, should work equally to empty dict `{}`
-        return
-
-    if not isinstance(value, dict):
-        raise ValidationError(
-            _("Given object is not a dictionary: %(value)r"),
-            params={"value": value},
-        )
